Make OidcRequestContextProperties modifiable
