+++ extensions/oidc-common/runtime/src/main/java/io/quarkus/oidc/common/OidcRequestContextProperties.java
@@ -19,22 +19,57 @@ public class OidcRequestContextProperties {
         this.properties = properties;
     }
 
+    /**
+     * Get property value
+     *
+     * @param name property name
+     * @return property value
+     */
     public <T> T get(String name) {
         @SuppressWarnings("unchecked")
         T value = (T) properties.get(name);
         return value;
     }
 
+    /**
+     * Get property value as String
+     *
+     * @param name property name
+     * @return property value as String
+     */
     public String getString(String name) {
         return (String) get(name);
     }
 
+    /**
+     * Get typed property value
+     *
+     * @param name property name
+     * @param type property type
+     * @return typed property value
+     */
     public <T> T get(String name, Class<T> type) {
         return type.cast(get(name));
     }
 
+    /**
+     * Get an unmodifiable view of the current context properties.
+     *
+     * @return all properties
+     */
     public Map<String, Object> getAll() {
         return Collections.unmodifiableMap(properties);
     }
 
+    /**
+     * Set the property
+     *
+     * @param name property name
+     * @param value property value
+     * @return this OidcRequestContextProperties instance
+     */
+    public OidcRequestContextProperties put(String name, Object value) {
+        properties.put(name, value);
+        return this;
+    }
 }
