+++ integration-tests/oidc-wiremock/src/main/resources/application.properties
@@ -244,8 +244,8 @@ quarkus.log.category."io.quarkus.oidc.runtime.OidcProviderClient".min-level=TRAC
 quarkus.log.category."io.quarkus.oidc.runtime.OidcProviderClient".level=TRACE
 quarkus.log.category."io.quarkus.it.keycloak.SignedUserInfoResponseFilter".min-level=TRACE
 quarkus.log.category."io.quarkus.it.keycloak.SignedUserInfoResponseFilter".level=TRACE
-quarkus.log.category."io.quarkus.it.keycloak.TokenResponseFilter".min-level=TRACE
-quarkus.log.category."io.quarkus.it.keycloak.TokenResponseFilter".level=TRACE
+quarkus.log.category."io.quarkus.it.keycloak.TokenRequestResponseFilter".min-level=TRACE
+quarkus.log.category."io.quarkus.it.keycloak.TokenRequestResponseFilter".level=TRACE
 quarkus.log.file.enable=true
 quarkus.log.file.format=%C - %s%n
